Make WSClient inherit from WSBaseClient
