+++ rpc/clients/typescript/src/ws_client.ts
@@ -5,6 +5,7 @@ import { BigNumber } from '@0x/utils';
 import { v4 as uuid } from 'uuid';
 import * as WebSocket from 'websocket';
 
+import { BaseWSClient } from './base_ws_client';
 import {
     AcceptedOrderInfo,
     ContractEvent,
@@ -34,14 +35,13 @@ import {
     ValidationResults,
     WSOpts,
 } from './types';
-import { WebSocketProvider } from './websocket_provider';
 
 const CLOSE_REASON_NO_HEARTBEAT = 3001;
 const CLOSE_DESCRIPTION_NO_HEARTBEAT = 'No heartbeat received';
 const CLEAR_SUBSCRIPTIONS_GRACE_PERIOD_MS = 100;
 
 const DEFAULT_RECONNECT_AFTER_MS = 5000;
-const DEFAULT_WS_OPTS = {
+const DEFAULT_WS_OPTS: WSOpts = {
     clientConfig: {
         // For some reason fragmenting the payloads causes the connection to close
         // Source: https://github.com/theturtle32/WebSocket-Node/issues/359
@@ -54,10 +54,9 @@ const DEFAULT_WS_OPTS = {
  * This class includes all the functionality related to interacting with a Mesh JSON RPC
  * websocket endpoint.
  */
-export class WSClient {
+export class WSClient extends BaseWSClient {
     private readonly _subscriptionIdToMeshSpecificId: ObjectMap<string>;
     private _heartbeatCheckIntervalId: number | undefined;
-    private readonly _wsProvider: WebSocketProvider;
     private static _convertRawAcceptedOrderInfos(rawAcceptedOrderInfos: RawAcceptedOrderInfo[]): AcceptedOrderInfo[] {
         const acceptedOrderInfos: AcceptedOrderInfo[] = [];
         rawAcceptedOrderInfos.forEach(rawAcceptedOrderInfo => {
@@ -190,14 +189,20 @@ export class WSClient {
      * @return  An instance of WSClient
      */
     constructor(url: string, wsOpts?: WSOpts) {
-        this._subscriptionIdToMeshSpecificId = {};
         if (wsOpts !== undefined && wsOpts.reconnectAfter === undefined) {
             wsOpts.reconnectAfter = DEFAULT_RECONNECT_AFTER_MS;
         }
-        this._wsProvider = createWebsocketProvider(
-            url,
-            wsOpts !== undefined ? (wsOpts as any) : DEFAULT_WS_OPTS,
-        );
+        const finalWSOpts = wsOpts !== undefined ? wsOpts : DEFAULT_WS_OPTS;
+        let connection: any;
+        // If running in Node.js environment
+        if (typeof process !== 'undefined' && process.versions != null && process.versions.node != null) {
+            const headers: any = finalWSOpts.headers || {};
+            connection = new (WebSocket.w3cwebsocket as any)(url, finalWSOpts.protocol, null, headers, null, finalWSOpts.clientConfig);
+        } else {
+            connection = new (window as any).WebSocket(url, finalWSOpts.protocol);
+        }
+        super(connection, finalWSOpts.reconnectAfter, finalWSOpts.timeout);
+        this._subscriptionIdToMeshSpecificId = {};
         // Intentional fire-and-forget
         // tslint:disable-next-line:no-floating-promises
         this._startInternalLivenessCheckAsync();
@@ -209,7 +214,7 @@ export class WSClient {
      */
     public async addOrdersAsync(signedOrders: SignedOrder[]): Promise<ValidationResults> {
         assert.isArray('signedOrders', signedOrders);
-        const rawValidationResults: RawValidationResults = await this._wsProvider.sendAsync('mesh_addOrders', [
+        const rawValidationResults: RawValidationResults = await this._sendAsync('mesh_addOrders', [
             signedOrders,
         ]);
         const validationResults: ValidationResults = {
@@ -240,7 +245,7 @@ export class WSClient {
         let snapshotID = ''; // New snapshot
 
         let page = 0;
-        const getOrdersResponse: GetOrdersResponse = await this._wsProvider.sendAsync('mesh_getOrders',
+        const getOrdersResponse: GetOrdersResponse = await this._sendAsync('mesh_getOrders',
             [
                 page,
                 perPage,
@@ -254,7 +259,7 @@ export class WSClient {
         do {
             rawOrderInfos = [...rawOrderInfos, ...ordersInfos];
             page++;
-            ordersInfos = (await this._wsProvider.sendAsync('mesh_getOrders', [page, perPage, snapshotID]))
+            ordersInfos = (await this._sendAsync('mesh_getOrders', [page, perPage, snapshotID]))
                 .ordersInfos;
         } while (Object.keys(ordersInfos).length > 0);
 
@@ -269,7 +274,7 @@ export class WSClient {
      */
     public async subscribeToOrdersAsync(cb: (orderEvents: OrderEvent[]) => void): Promise<string> {
         assert.isFunction('cb', cb);
-        const orderEventsSubscriptionId = await this._wsProvider.subscribeAsync('mesh_subscribe', 'orders', []);
+        const orderEventsSubscriptionId = await this._subscribeAsync('mesh_subscribe', 'orders', []);
         const id = uuid();
         this._subscriptionIdToMeshSpecificId[id] = orderEventsSubscriptionId;
 
@@ -289,7 +294,7 @@ export class WSClient {
             });
             cb(orderEvents);
         };
-        this._wsProvider.on(orderEventsSubscriptionId, orderEventsCallback as any);
+        this.on(orderEventsSubscriptionId, orderEventsCallback as any);
         return id;
     }
     /**
@@ -302,7 +307,7 @@ export class WSClient {
         if (meshSubscriptionId === undefined) {
             throw new Error(`Subscription not found with ID: ${subscriptionId}`);
         }
-        await this._wsProvider.unsubscribeAsync(meshSubscriptionId, 'mesh_unsubscribe');
+        await this._unsubscribeAsync(meshSubscriptionId, 'mesh_unsubscribe');
     }
     /**
      * Get notified when the underlying WS connection closes normally. If it closes with an
@@ -310,7 +315,7 @@ export class WSClient {
      * @param cb callback to call when WS connection closes
      */
     public onClose(cb: () => void): void {
-        this._wsProvider.on('close', () => {
+        this.on('close', () => {
             cb();
         });
     }
@@ -319,7 +324,7 @@ export class WSClient {
      * @param cb callback to call with the error when it occurs
      */
     public onReconnected(cb: () => void): void {
-        this._wsProvider.on('reconnected', () => {
+        this.on('reconnected', () => {
             cb();
         });
     }
@@ -331,12 +336,12 @@ export class WSClient {
         clearInterval(this._heartbeatCheckIntervalId);
         // HACK(fabio): We fire-and-forget the call to clear subscriptions since we don't want `destroyAsync()`
         // to block if the connection is having issues. The hacky part though is that we will get an error if
-        // we try to send a payload on a connection _we know_ is closed. We therefore need to call `disconnect`
+        // we try to send a payload on a connection _we know_ is closed. We therefore need to call `_disconnect`
         // after a timeout so that we are sure we've already attempted to send the unsubscription payloads before
         // we forcefully close the connection
-        this._wsProvider.clearSubscriptionsAsync('mesh_unsubscribe');
+        this._clearSubscriptionsAsync('mesh_unsubscribe');
         await new Promise<NodeJS.Timer>(resolve => setTimeout(resolve, CLEAR_SUBSCRIPTIONS_GRACE_PERIOD_MS));
-        this._wsProvider.disconnect(WebSocket.connection.CLOSE_REASON_NORMAL, 'Normal connection closure');
+        this._disconnect(WebSocket.connection.CLOSE_REASON_NORMAL, 'Normal connection closure');
     }
     /**
      * Subscribe to the 'heartbeat' topic and receive an ack from the Mesh every 5 seconds. This method
@@ -346,7 +351,7 @@ export class WSClient {
      */
     private async _subscribeToHeartbeatAsync(cb: (ack: string) => void): Promise<string> {
         assert.isFunction('cb', cb);
-        const heartbeatSubscriptionId = await this._wsProvider.subscribeAsync('mesh_subscribe', 'heartbeat', []);
+        const heartbeatSubscriptionId = await this._subscribeAsync('mesh_subscribe', 'heartbeat', []);
         const id = uuid();
         this._subscriptionIdToMeshSpecificId[id] = heartbeatSubscriptionId;
 
@@ -354,7 +359,7 @@ export class WSClient {
             this._subscriptionIdToMeshSpecificId[id] = eventPayload.subscription;
             cb(eventPayload.result);
         };
-        this._wsProvider.on(heartbeatSubscriptionId, orderEventsCallback as any);
+        this.on(heartbeatSubscriptionId, orderEventsCallback as any);
         return id;
     }
     private async _startInternalLivenessCheckAsync(): Promise<void> {
@@ -374,38 +379,11 @@ export class WSClient {
                 lastHeartbeatTimestampMs + twentySecondsInMs < new Date().getTime();
             if (haveTwentySecondsPastWithoutAHeartBeat) {
                 // If connected, we haven't received a heartbeat in over 20 seconds, re-connect
-                if (this._wsProvider.isConnected()) {
-                    this._wsProvider.disconnect(CLOSE_REASON_NO_HEARTBEAT, CLOSE_DESCRIPTION_NO_HEARTBEAT);
+                if (this._isConnected()) {
+                    this._disconnect(CLOSE_REASON_NO_HEARTBEAT, CLOSE_DESCRIPTION_NO_HEARTBEAT);
                 }
                 lastHeartbeatTimestampMs = new Date().getTime();
             }
         }, oneSecondInMs) as any;
     }
 }
-
-/**
- * Return an WebsocketProvider object
- * @param url WebSocket endpoint
- * @param options WebSocket options
- * @returns A WebSocketProvider instance
- */
-function createWebsocketProvider(url: string, options: WSOpts = {}): WebSocketProvider {
-        let connection = '';
-
-        // runtime is of type node
-        if (typeof process !== 'undefined' && process.versions != null && process.versions.node != null) {
-            const headers: any = options.headers || {};
-            const urlObject = new URL(url);
-
-            if (!headers.authorization && urlObject.username && urlObject.password) {
-                const authToken = Buffer.from(`${urlObject.username}:${urlObject.password}`).toString('base64');
-                headers.authorization = `Basic ${authToken}`;
-            }
-
-            connection = new (WebSocket.w3cwebsocket as any)(url, options.protocol, null, headers, null, options.clientConfig);
-        } else {
-            connection = new (window as any).WebSocket(url, options.protocol);
-        }
-
-        return new WebSocketProvider(connection, options.reconnectAfter, options.timeout);
-    }
